fix : 게임 결과와 게임 결과 계산을 분리 및 테스트 코드 작성
